amazonka: Extract Auth exceptions into separate module
